pyproject.toml
[project]
name = "bdsys-ec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.0.2",
    "apache-airflow-providers-apache-spark>=5.3.0",
    "flask-appbuilder>=4.7.0",
    "ipykernel>=6.29.5",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pendulum>=3.1.0",
    "pyarrow>=20.0.0",
    "pyspark>=4.0.0",
    "seaborn>=0.13.2",
]
